修改：缓存命中率从整数改为浮点数并保留一位小数显示，提升前端 Context 条精度

+++ soloncode-cli/src/main/java/org/noear/solon/codecli/portal/web/event/payload/SystemContextPayload.java
@@ -15,6 +15,5 @@ public class SystemContextPayload implements Serializable {
     private Integer tokens;
     private Integer count;
     private Integer contextLimit;
-    /** 缓存命中率（%），驱动前端 Context 条的 "Cache: N%" 指示。 */
-    private Integer cacheRate;
+    private Double cacheRate;
 }
